+++ java/yb-cdc/src/test/java/org/yb/cdc/TestDBStreamInfo.java
@@ -29,7 +29,7 @@ import org.yb.client.TestUtils;
 import org.yb.client.YBClient;
 import org.yb.client.YBTable;
 import org.yb.master.MasterDdlOuterClass.ListTablesResponsePB.TableInfo;
-import org.yb.util.YBTestRunnerNonTsanOnly;
+import org.yb.YBTestRunner;
 
 import static org.yb.AssertionWrappers.*;
 
@@ -39,7 +39,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-@RunWith(value = YBTestRunnerNonTsanOnly.class)
+@RunWith(value = YBTestRunner.class)
 public class TestDBStreamInfo extends CDCBaseClass {
   private final static Logger LOG = LoggerFactory.getLogger(TestDBStreamInfo.class);