+++ java/yb-cdc/src/test/java/org/yb/cdc/ysql/TestDDL.java
@@ -23,14 +23,14 @@ import org.yb.cdc.common.*;
 import org.yb.cdc.util.CDCSubscriber;
 
 import org.yb.cdc.CdcService.RowMessage.Op;
-import org.yb.util.YBTestRunnerNonTsanOnly;
+import org.yb.YBTestRunner;
 
 import static org.yb.AssertionWrappers.*;
 
 import java.util.ArrayList;
 import java.util.List;
 
-@RunWith(value = YBTestRunnerNonTsanOnly.class)
+@RunWith(value = YBTestRunner.class)
 public class TestDDL extends CDCBaseClass {
   private final static Logger LOG = LoggerFactory.getLogger(TestDDL.class);