+++ java/yb-pgsql/src/test/java/org/yb/pgsql/TestBatchCopyFrom.java
@@ -33,9 +33,9 @@ import com.yugabyte.core.BaseConnection;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.yb.client.TestUtils;
-import org.yb.util.YBTestRunnerNonTsanOnly;
+import org.yb.YBTestRunner;
 
-@RunWith(value = YBTestRunnerNonTsanOnly.class)
+@RunWith(value = YBTestRunner.class)
 public class TestBatchCopyFrom extends BasePgSQLTest {
   private static final Logger LOG = LoggerFactory.getLogger(TestBatchCopyFrom.class);