+++ java/yb-pgsql/src/test/java/org/yb/pgsql/TestMasterMetrics.java
@@ -26,12 +26,12 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.yb.minicluster.Metrics;
 import org.yb.minicluster.MiniYBDaemon;
-import org.yb.util.YBTestRunnerNonTsanOnly;
+import org.yb.YBTestRunner;
 
 /**
  * Runs the pg_regress test suite on YB code.
  */
-@RunWith(value = YBTestRunnerNonTsanOnly.class)
+@RunWith(value = YBTestRunner.class)
 public class TestMasterMetrics extends BasePgSQLTest {
   private static final Logger LOG = LoggerFactory.getLogger(TestMasterMetrics.class);