+++ java/yb-pgsql/src/test/java/org/yb/pgsql/TestPgParallelSelect.java
@@ -18,14 +18,14 @@ import org.junit.runner.RunWith;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.yb.util.BuildTypeUtil;
-import org.yb.util.YBTestRunnerNonTsanOnly;
+import org.yb.YBTestRunner;
 
 import java.sql.Statement;
 import java.util.Map;
 
 import static org.yb.AssertionWrappers.*;
 
-@RunWith(value=YBTestRunnerNonTsanOnly.class)
+@RunWith(value=YBTestRunner.class)
 public class TestPgParallelSelect extends BasePgSQLTest {
   private static final Logger LOG = LoggerFactory.getLogger(TestPgParallelSelect.class);
   private static final int kNumShardsPerTserver = 20;