+++ java/yb-pgsql/src/test/java/org/yb/pgsql/TestPgSelect.java
@@ -21,9 +21,9 @@ import org.slf4j.LoggerFactory;
 import org.yb.minicluster.RocksDBMetrics;
 
 import org.yb.util.BuildTypeUtil;
-import org.yb.util.YBTestRunnerNonTsanOnly;
+import org.yb.YBTestRunner;
 import org.yb.util.RegexMatcher;
-import org.yb.util.YBTestRunnerNonTsanOnly;
+import org.yb.YBTestRunner;
 
 import java.math.BigDecimal;
 import java.sql.Connection;
@@ -42,7 +42,7 @@ import java.util.stream.Collectors;
 
 import static org.yb.AssertionWrappers.*;
 
-@RunWith(value=YBTestRunnerNonTsanOnly.class)
+@RunWith(value=YBTestRunner.class)
 public class TestPgSelect extends BasePgSQLTest {
   private static final Logger LOG = LoggerFactory.getLogger(TestPgSelect.class);