+++ managed/ui/src/redesign/features/universe/universe-form/form/fields/AssignPublicIPField/AssignPublicIPField.tsx
@@ -2,9 +2,10 @@ import React, { ReactElement } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useFormContext } from 'react-hook-form';
 import { Box } from '@material-ui/core';
-import { YBHelper, YBLabel, YBToggleField } from '../../../../../../components';
+import { YBLabel, YBToggleField, YBTooltip } from '../../../../../../components';
 import { UniverseFormData } from '../../../utils/dto';
 import { ASSIGN_PUBLIC_IP_FIELD } from '../../../utils/constants';
+import InfoMessageIcon from '../../../../../../assets/info-message.svg';
 
 interface AssignPublicIPFieldProps {
   disabled: boolean;
@@ -13,24 +14,34 @@ interface AssignPublicIPFieldProps {
 export const AssignPublicIPField = ({ disabled }: AssignPublicIPFieldProps): ReactElement => {
   const { control } = useFormContext<UniverseFormData>();
   const { t } = useTranslation();
+  const assignIPTooltipText = t('universeForm.instanceConfig.assignPublicIPHelper');
 
   return (
-    <Box display="flex" width="100%" data-testid="AssignPublicIPField-Container">
-      <YBLabel dataTestId="AssignPublicIPField-Label">
-        {t('universeForm.instanceConfig.assignPublicIP')}
-      </YBLabel>
-      <Box flex={1}>
-        <YBToggleField
-          name={ASSIGN_PUBLIC_IP_FIELD}
-          inputProps={{
-            'data-testid': 'AssignPublicIPField-Toggle'
-          }}
-          control={control}
-          disabled={disabled}
-        />
-        <YBHelper dataTestId="AssignPublicIPField-Helper">
-          {t('universeForm.instanceConfig.assignPublicIPHelper')}
-        </YBHelper>
+    <Box
+      display="flex"
+      width="100%"
+      data-testid="AssignPublicIPField-Container"
+      mt={2}
+      mb={2}
+      ml={2}
+    >
+      <YBToggleField
+        name={ASSIGN_PUBLIC_IP_FIELD}
+        inputProps={{
+          'data-testid': 'AssignPublicIPField-Toggle'
+        }}
+        control={control}
+        disabled={disabled}
+      />
+
+      <Box flex={1} alignSelf="center">
+        <YBLabel dataTestId="AssignPublicIPField-Label">
+          {t('universeForm.instanceConfig.assignPublicIP')}
+          &nbsp;
+          <YBTooltip title={assignIPTooltipText}>
+            <img alt="Info" src={InfoMessageIcon} />
+          </YBTooltip>
+        </YBLabel>
       </Box>
     </Box>
   );