+++ managed/ui/src/redesign/features/universe/universe-form/form/fields/SystemdField/SystemdField.tsx
@@ -16,7 +16,7 @@ export const SystemDField = ({ disabled }: SystemDFieldProps): ReactElement => {
 
   return (
     <Box display="flex" width="100%" data-testid="SystemDField-Container">
-      <YBLabel dataTestId="SystemDField-Label">
+      <YBLabel dataTestId="SystemDField-Label" width="224px">
         {t('universeForm.advancedConfig.enableSystemD')}
       </YBLabel>
       <Box flex={1}>