+++ managed/ui/src/utils/UniverseUtils.js
@@ -100,6 +100,36 @@ export const getUniverseNodeCount = (nodeDetailsSet, cluster = null) => {
   ).length;
 };
 
+export const getUniverseDedicatedNodeCount = (nodeDetailsSet, cluster = null) => {
+  const nodes = nodeDetailsSet ?? [];
+  const numTserverNodes = nodes.filter(
+    (node) =>
+      (cluster === null || node.placementUuid === cluster.uuid) &&
+      _.includes(nodeInClusterStates, node.state) &&
+      node.dedicatedTo === 'TSERVER'
+  ).length;
+  const numMasterNodes = nodes.filter(
+    (node) =>
+      (cluster === null || node.placementUuid === cluster.uuid) &&
+      _.includes(nodeInClusterStates, node.state) &&
+      node.dedicatedTo === 'MASTER'
+  ).length;
+  return {
+    numTserverNodes,
+    numMasterNodes
+  };
+};
+
+export const isDedicatedNodePlacement = (currentUniverse) => {
+  let isDedicatedNodes = false;
+  if (!currentUniverse?.universeDetails) return isDedicatedNodes;
+
+  const clusters = currentUniverse.universeDetails.clusters;
+  const primaryCluster = clusters && getPrimaryCluster(clusters);
+  isDedicatedNodes = primaryCluster.userIntent.dedicatedNodes;
+  return isDedicatedNodes;
+};
+
 export function getProviderMetadata(provider) {
   return PROVIDER_TYPES.find((providerType) => providerType.code === provider.code);
 }
@@ -139,7 +169,7 @@ export function isKubernetesUniverse(currentUniverse) {
     isDefinedNotNull(currentUniverse.universeDetails) &&
     isDefinedNotNull(getPrimaryCluster(currentUniverse.universeDetails.clusters)) &&
     getPrimaryCluster(currentUniverse.universeDetails.clusters).userIntent.providerType ===
-    'kubernetes'
+      'kubernetes'
   );
 }
 
@@ -174,7 +204,11 @@ export const isOnpremUniverse = (universe) => {
 };
 
 export const isPausableUniverse = (universe) => {
-  return isUniverseType(universe, 'aws') || isUniverseType(universe, 'gcp') || isUniverseType(universe, 'azu');
+  return (
+    isUniverseType(universe, 'aws') ||
+    isUniverseType(universe, 'gcp') ||
+    isUniverseType(universe, 'azu')
+  );
 };
 
 // Reads file and passes content into Promise.resolve