+++ src/yb/master/catalog_manager_util.h
@@ -176,16 +176,6 @@ class CatalogManagerUtil {
       const Schema& schema, uint32_t schema_version, const PartitionSchema& partition_schema,
       tablet::TableInfoPB* pb);
 
-  // If the table is already hidden or deleted then treats it as a duplicate delete request.
-  // Don't invoke this function if attempting to hard delete an already hidden table.
-  static bool IsDuplicateDeleteTableRequest(const scoped_refptr<TableInfo>& table) {
-    if (!table->IsOperationalForClient()) {
-      LOG(INFO) << "Table " << table->id() << " is already deleted. Duplicate request.";
-      return true;
-    }
-    return false;
-  }
-
  private:
   CatalogManagerUtil();